+++ Assets/Game/101. Scripts/DummyDamageHandler.cs
@@ -3,7 +3,7 @@ using UnityEngine;
 
 class DummyDamageHandler : DamageHandler
 {
-    public override void DamageHandle(int damage)
+    public override void DamageHandle(double damage)
     {
         Debug.Log($"앗 너무 아파요 ㅠㅠ {damage}");
     }
